Leverages the search filter on services
Protobuf services are now filtered for using "ServiceEntity" as a suffix. This filtering permits non-entity Protobuf service declarations consistent in naming with our examples. As we introduce stronger declarations of what constitutes entities etc. in Protobuf files, we can widen this filter for all services. Also, we can be even more explicit in the name of a service entity by expressing a regex, particularly in the case where the user does not like our naming convention.