Removes the use of javac
The removal of javac is necessary so that this library may be built using GraalVM Native and used outside of where a JDK exists. The code also becomes dramatically simpler as we now pass in a protobuf descriptor rather than having to compile classes to obtain one. Clients of this library can call require that protoc generates the descriptor via the --descriptor_set_out option.