Provides a native image CLI for JavaScript code generation
Source generators for JavaScript have been introduced, albeit non-functioning (that's for a separate PR). This PR introduces a native image CLI for JavaScript code generation using GraalVM. The CLI behaves in a similar fashion to the akkasls-maven-plugin.

One breaking change is that the generator for Java has been moved into a `com.lightbend.akkasls.codegen.java` package i.e. the `java` is another level down. This is so it is viable to depend on multiple code generator libraries here, although note that there is no immedate requirement for this.