Substitute the archetype version for the plugin
The plugin's version of the generated pom.xml was previously using the project's version, and not the archetype's version. By default, the plugin version should be the same as the archetype version so we introduce a new property for the plugin's version. The default of this property is the archetype's version.