Corrected proto desc usage
For JS codegen, we must provide the list of proto files that contain the service describing the event sourced entity. Prior to this commit, I had assumed that the protodesc was required. As a result of the changes here,  we now allow the protobuf source dir to be passed in. We can then express the proto source files in relation to the js source files.