Improved error reporting
Fixes #XXX

This greatly improves the error reporting from the proxy to the user
function. It adds the following info to the report error message:

* code - A code that we can use for many things, including metrics,
  unambiguously referring to classes of errors, etc.
* detail - A multi-line detail field that explains to the user how they
  might fix the error that they have made.
* sourceLocations - Zero to many locations in the protobuf sources that
  were the source of the error.

Not every error we report has been switched to taking full advantage of
this, but a few have.

To support this, we needed to ensure the user functions include source
code info in their returned protobuf descriptors. This is trivial in
JavaScript, a little more involved in Java and will require additional
Maven configuration, but I've put some very helpful help to instruct a
user how to configure it if they haven't configured it correctly.