test: eventing catchOthers mismatch
* test: eventing catchOthers mismatch

Reproducer for the scenario encountered with the wirelessmesh example described
in #XXX: the incoming protobuf message is not of any of the recognized
types.

`Event2Proto.toAny` just wraps the incoming `CustomerLocationAdded` into
an `Any` without looking at whether `CustomerLocationAdded` is actually
supported by any of the action methods. The router then (correctly)
routes the message to the method accepting protobuf `Any`.

The SDK then receives the `Any`, and interprets its contents as an
`Any`, consistent with any other methods.

We have to decide whether we want to add logic to the server to wrap the
payload into an additional `Any` when passing objects to the catchall
method, or we want to add logic to the SDK to pass the `Any` as-is to
the method when the `typeUrl` is different from
`types.googleapis.com/google.protobuf.Any`.

* fix(proxy): wrap catch-all events in Any

This makes the API between the proxy and the user function consistent
between calling functions that take an `Any` 'regularly' and calling
functions that take an `Any` as 'catch-others' function. As such it
complicates the proxy a bit, but keeps the API and SDK's simpler,
and keeps the 'fall back to catch-all' logic in one place.

* scalafmt