First pass using codegen instead of reflection
Manually create a `CustomerValueEntityHandler` that would be generated
as part of codegen.

Added basic code to hook it into `Main`, but that API would still change
and partly be generated.

Reading the codegen code, it's clear we're losing some features that
used to be available with codegen, marked them with TODO's - we should
consider whether we're OK with losing them or want to introduce new
mechanisms.