feat: basic sbt plugin (#384)
With dummy codegen.

This provides nice infrastructure for the part of codegen that happens each time the .proto changes, but not yet for the 'one-off' code generation of code that is supposed to be adapted by the user later. Perhaps that should be an explicit command anyway, to prevent 'leaving behind' generated code and confusing people when class or package names change.