feat: sbt plugin 'unmanaged' codegen
Not entirely happy with the approach: generating the code from the regular
protoc generators, but then 'filtering them out' of the managed sources
(that they get added to automatically) afterwards, and copying them over
'manually'.

It does save us from duplicating part of the sbt-protoc implementation
(in particular finding the correct `.proto` files and converting them
into `com.google.protobuf.Descriptors.FileDescriptor` instances).

Let's try it out, but if it leads to trouble we should definitely consider
avoiding the generators and doing that step 'ourselves'.