docs: small improvements (#1682)
* Correct incomplete statements in Views.

The sentence in Creating a View from an Event Sourced Entity, and a similar sentence in Creating a View from a Value Entity were not holding.

* Remove duplicate word.

* Correct naming in ACL section.

In section [Configuring ACLs](https://docs.kalix.io/java/access-control.html#_configuring_acls) there is a reference to `my-service`, while in the code example is referred to as `service-b`. Changed code to use name `service-b`.